refactor: Deprecate and remove unused chat system components
- Move deprecated chain.py, think.py, and task.py to deprecated directory
- Update main.py import to use deprecated chain module
- Remove unused tools and modules related to previous chat system architecture
- Simplify project structure by removing redundant files
