+++ chat/deprecated/think.py
